本地令牌界面 view层结构更改

+++ System.Application.SteamTools.Client.Desktop/Services/Implementation/SteamServiceImpl.cs
@@ -52,7 +52,8 @@ namespace System.Application.Services.Implementation
             {
                 if (steamProcess.Contains(p.ProcessName, StringComparer.OrdinalIgnoreCase))
                 {
-                    p.Kill();
+                    if (!p.CloseMainWindow())
+                        p.Kill();
                 }
             }
         }
